requirements.txt
# Core Dependencies for Finance HMM API
# Install with: pip install -r requirements.txt

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Data Science & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Hidden Markov Model
hmmlearn>=0.3.0

# Financial Data
yfinance>=0.2.28

# Visualization
matplotlib>=3.7.0

# AI/LLM Agent
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40
tavily-python>=0.3.0

# Environment & Utilities
python-dotenv>=1.0.0
requests>=2.31.0
